% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_et0.R
\name{compute_et0}
\alias{compute_et0}
\title{Compute ET0 locally}
\usage{
compute_et0(
  date,
  lat,
  elev,
  t_max,
  t_min,
  ws_mean,
  sr_mean,
  rh_max = NULL,
  rh_min = NULL,
  rh_mean = NULL,
  ws_h = 2
)
}
\arguments{
\item{date}{Date or str. Date in YYYY-MM-DD format (Mandatory)}

\item{lat}{numeric. Latitude in decimal degrees (Mandatory)}

\item{elev}{numeric. Elevation in meters above sea level (Mandatory)}

\item{t_max}{numeric. Maximum temperature in degrees Celsius (Mandatory)}

\item{t_min}{numeric. Minimum temperature in degrees Celsius (Mandatory)}

\item{ws_mean}{numeric. Mean wind speed in m/s (Mandatory)}

\item{sr_mean}{numeric. Mean solar radiation in W/m^2/day (Mandatory)}

\item{rh_max}{numeric. Maximum relative humidity in percentage (Optional)}

\item{rh_min}{numeric. Minimum relative humidity in percentage (Optional)}

\item{rh_mean}{numeric. Mean relative humidity in percentage (Optional)}

\item{ws_h}{numeric. Height of the wind speed measurement in meters (Optional, default = 2)}
}
\value{
A numeric value with the reference evapotranspiration (ET0) in mm/day.
}
\description{
This function computes the reference evapotranspiration (ET0) using the Penman-Monteith method. It's based on "Step by step calculation of the Penman-Monteith Evapotranspiration (FAO-56 Method)" by H. Allen, R. Pereira, D. Raes, and M. Smith.
}
\details{
There are 4 different ways to calculate the vapor pressure (ea) depending on the input data:
\enumerate{
\item If both rh_max and rh_min are provided, ea is calculated as the average of the vapor pressure at the maximum and minimum temperature.
\item If only rh_max is provided, ea is calculated as the vapor pressure at the minimum temperature.
\item If only rh_mean is provided, ea is calculated as the vapor pressure at the mean temperature.
\item If none of the relative humidity values are provided, ea is calculated as the vapor pressure at the mean temperature.
}
}
\examples{
compute_et0(date='2024-03-09',lat=-30.631261,elev=80,t_min=7.91,t_max=27.69,
            rh_min=37.71,rh_max=93.2,ws_mean=1.364,sr_mean=272.824
            ) # WS Software ET0: 4.32 mm
compute_et0(date='2024-03-09',lat=-30.631261,elev=80,t_min=7.91,t_max=27.69,
            rh_max=93.2,ws_mean=1.364,sr_mean=272.824
            )# WS Software ET0: 4.32 mm
compute_et0(date='2024-03-09',lat=-30.631261,elev=80,t_min=7.91,t_max=27.69,
            rh_mean=66.81,ws_mean=1.364,sr_mean=272.824
            )# WS Software ET0: 4.32 mm
compute_et0(date='2024-03-09',lat=-30.631261,elev=80,t_min=7.91,t_max=27.69,
            ws_mean=1.364,sr_mean=272.824
            )# WS Software ET0: 4.32 mm
compute_et0(date='2024-03-02',lat=-30.592033,elev=292,t_min=11.82,
            t_max=26.72,rh_min=51.18,rh_max=94.1,ws_mean=2.471,
            sr_mean=259.846, ws_h = 5)# WS Software ET0: 4.29 mm

}
