% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et0.R
\name{et0}
\alias{et0}
\title{Compute ET0}
\usage{
et0(
  lat,
  lon,
  elev,
  month,
  t_max,
  t_min,
  rh_max,
  rh_min,
  ws_mean,
  sr_mean,
  debug = FALSE,
  user = NULL
)
}
\arguments{
\item{lat}{numeric, latitude in decimal degrees. Must be between -90 and 90. Mandatory}

\item{lon}{numeric, longitude in decimal degrees. Must be between -180 and 180. Mandatory}

\item{elev}{numeric, elevation in meters above sea level. Mandatory}

\item{month}{integer or vector, month of the year. Mandatory}

\item{t_max}{numeric or vector, maximum temperature in degrees Celsius. Mandatory}

\item{t_min}{numeric or vector, minimum temperature in degrees Celsius. Mandatory}

\item{rh_max}{numeric or vector, maximum relative humidity in percentage. Mandatory}

\item{rh_min}{numeric or vector, minimum relative humidity in percentage. Mandatory}

\item{ws_mean}{numeric or vector, mean wind speed in meters per second. Mandatory}

\item{sr_mean}{numeric or vector, mean solar radiation in Watts per square meter. Mandatory}

\item{debug}{logical, print debug information}

\item{user}{character, user email}
}
\value{
numeric value of ET0 in mm/day
}
\description{
Compute evapotranspiration using the FAO-56 Penman-Monteith method
}
\examples{
\dontrun{
et0(lat = -30, lon = -70, elev = 20,
    month = 1, t_max = 29.6, t_min = 22.8,
    rh_max = 90, rh_min = 70, ws_mean = 0.2,
    sr_mean = 200)

et0(lat = -30, lon = -70, elev = 20,
    month = c(1,2), t_max = c(29.6, 30), t_min = c(22.8, 21),
    rh_max = c(90, 80), rh_min = c(70, 60), ws_mean = c(0.2, 0.9),
    sr_mean = c(200, 180))
}

}
