% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSensorsData.R
\name{getSensorData}
\alias{getSensorData}
\title{Get data series from an specific sensor}
\usage{
getSensorData(
  s_cod,
  start_date,
  end_date,
  user = NULL,
  interv = NULL,
  header = NULL,
  nan_value = NULL,
  output = NULL,
  number_format = NULL,
  filename = NULL
)
}
\arguments{
\item{s_cod}{int or string, Station code, mandatory field}

\item{start_date}{string, Start date in YYYY-MM-DD or YYYY-MM-DD HH:MM:SS format, mandatory field}

\item{end_date}{string, End date in YYYY-MM-DD or YYYY-MM-DD HH:MM:SS format, mandatory field}

\item{user}{string, user name, add your email or user name (default: NULL)}

\item{interv}{string, Time interval, it must be a subset of c("hora"/hour, "dia"/day, "mes"/month) (default: NULL)}

\item{header}{boolean, omit header? (default: NULL)}

\item{nan_value}{string or int, value to replace NaN (default: NULL)}

\item{output}{string, output format, it must be a subset of c("", "json") (default: NULL, equvalent to "")}

\item{number_format}{string, number and decimal length, in formant "n.d" like "10.3" (default: NULL)}

\item{filename}{string, path to save the data frame as a CSV file (default: NULL)}
}
\value{
a data frame with the sensor data
}
\description{
This functions retrieves the data series from a specific sensor in a specific time interval. The output is a data frame, although if json is asked, extra information is showed
}
\examples{
\dontrun{
getSensorData(
  s_cod = "RPLTA", start_date = "2012-08-20",
  end_date = "2012-08-27", user = "anon@host.com"
)
getSensorData(
  s_cod = "RPLTA", start_date = "2019-05-13 12:00:00",
  end_date = "2019-05-13 15:00:00", user = "anon@host.com"
)
getSensorData(
  s_cod = "RPLTA", start_date = "2012-08-20",
  end_date = "2012-08-27", interv = "dia"
)
getSensorData(
  s_cod = "RPLTA", start_date = "2016-08-20",
  end_date = "2016-08-27", user = "anon@host.com", nan_value = 99999
)
getSensorData(
  s_cod = 29, start_date = "2015-08-01",
  end_date = "2015-08-02", nan_value = "", number_format = "10.3"
)
getSensorData(
  s_cod = 29, start_date = "2015-08-01",
  end_date = "2015-08-02", output = "json"
)
}

}
