% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSensorsList.R
\name{getSensorsList}
\alias{getSensorsList}
\title{Get sensors list}
\usage{
getSensorsList(
  p_cod = "ceazamet",
  e_cod,
  user = NULL,
  tm_cod = NULL,
  cols = NULL,
  filename = NULL
)
}
\arguments{
\item{p_cod}{string, The code of the provider, currently only "ceazamet" is supported}

\item{e_cod}{int or string, Station code, mandatory field}

\item{user}{string, user name, add your email or user name (default: NULL)}

\item{tm_cod}{string, type of measurement, it must be a subset of c("ta_c", "hr", "vv_ms", "dv", "rs_w", "pp_mm") (default: NULL)}

\item{cols}{string or vector of strings, Columns to retrieve, it must be a subset of c("e_cod", "s_cod", "tf_nombre", "um_notacion", "s_altura", "s_ultima_lectura", "tm_cod", "s_primera_lectura") (default: NULL)}

\item{filename}{string, path to save the data frame as a CSV file (default: NULL)}
}
\value{
A data frame with the sensors list
}
\description{
This function retrieves the list of sensors for a given station, use \code{e_cod} from \code{getStationsList} to get the appropriate code for the station.
}
\examples{
\dontrun{
getSensorsList(e_cod = 9, user = "anon@host.com")
getSensorsList(
  e_cod = 6, user = "anon@host.com",
  cols = c(
    "e_cod", "s_cod", "tf_nombre", "um_notacion",
    "s_altura", "s_ultima_lectura"
  )
)
getSensorsList(e_cod = 6, user = "anon@host.com", tm_cod = "ta_c")
getSensorsList(
  e_cod = 6, user = "anon@host.com",
  cols = c("e_cod", "s_cod", "tm_cod", "s_altura")
)
}

}
